requirements.txt
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
fastavro==1.12.1
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
idna==3.11
numpy==2.4.0
pandas==2.3.3
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
